# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  request_schema.CreateRequest:
    properties:
      endpoint_path:
        type: string
      max_tokens_key:
        type: string
      name:
        type: string
      provider_id:
        type: string
      request_template:
        type: string
      response_image_path:
        type: string
      response_text_path:
        type: string
      supports_streaming:
        type: boolean
      supports_temperature:
        type: boolean
      system_role_key:
        type: string
    type: object
  request_schema.RequestSchema:
    properties:
      created_at:
        type: string
      endpoint_path:
        type: string
      id:
        type: string
      max_tokens_key:
        type: string
      name:
        type: string
      provider_id:
        type: string
      request_template:
        type: string
      response_image_path:
        type: string
      response_text_path:
        type: string
      supports_streaming:
        type: boolean
      supports_temperature:
        type: boolean
      system_role_key:
        type: string
      updated_at:
        type: string
    type: object
  request_schema.UpdateRequest:
    properties:
      endpoint_path:
        type: string
      max_tokens_key:
        type: string
      name:
        type: string
      request_template:
        type: string
      response_image_path:
        type: string
      response_text_path:
        type: string
      supports_streaming:
        type: boolean
      supports_temperature:
        type: boolean
      system_role_key:
        type: string
    type: object
  sitemap.HistoryResponse:
    properties:
      baseURL:
        type: string
      createdAt:
        type: string
      id:
        type: string
      productId:
        type: string
      productName:
        type: string
      sitemapURL:
        type: string
      status:
        description: success, failed, pending
        type: string
      title:
        type: string
      totalURLs:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@seo-backend.com
    name: API Support
  description: API Documentation for SEO Backend Service
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SEO Backend API
  version: "1.0"
paths:
...
  /api/request-schemas:
    get:
      consumes:
      - application/json
      description: Get all request schemas with pagination
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of request schemas with pagination
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all request schemas
      tags:
      - Request Schemas
    post:
      consumes:
      - application/json
      description: Create a new request schema for API validation
      parameters:
      - description: Request schema creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_schema.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Request schema created
          schema:
            $ref: '#/definitions/request_schema.RequestSchema'
        "400":
          description: Bad request - invalid provider_id, name, or endpoint_path
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Duplicate entry
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new request schema
      tags:
      - Request Schemas
  /api/request-schemas/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a request schema by ID
      parameters:
      - description: Request Schema ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: deleted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Request schema not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a request schema
      tags:
      - Request Schemas
    get:
      consumes:
      - application/json
      description: Get a specific request schema by its ID
      parameters:
      - description: Request Schema ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Request schema details
          schema:
            $ref: '#/definitions/request_schema.RequestSchema'
        "404":
          description: Request schema not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get request schema by ID
      tags:
      - Request Schemas
    put:
      consumes:
      - application/json
      description: Update an existing request schema by ID
      parameters:
      - description: Request Schema ID
        in: path
        name: id
        required: true
        type: string
      - description: Request schema update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_schema.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated request schema
          schema:
            $ref: '#/definitions/request_schema.RequestSchema'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Request schema not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Duplicate entry
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a request schema
      tags:
      - Request Schemas
  /api/request-schemas/provider/{provider_id}:
    get:
      consumes:
      - application/json
      description: Get all request schemas for a specific provider
      parameters:
      - description: Provider ID
        in: path
        name: provider_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of request schemas
          schema:
            items:
              $ref: '#/definitions/request_schema.RequestSchema'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get request schemas by provider
      tags:
      - Request Schemas
  /api/sitemap:
    get:
      consumes:
      - application/json
      description: Generate an XML sitemap for a product or base URL
      parameters:
      - description: Product ID to generate sitemap for
        in: query
        name: product_id
        type: string
      - description: Base URL for the sitemap
        in: query
        name: base_url
        required: true
        type: string
      - default: false
        description: Include images in sitemap
        in: query
        name: include_images
        type: boolean
      produces:
      - text/xml
      responses:
        "200":
          description: XML Sitemap
          headers:
            X-Base-URL:
              description: Base URL used for generation
              type: string
            X-Generated-At:
              description: Timestamp when sitemap was generated
              type: string
            X-Product-ID:
              description: Product ID used for generation
              type: string
            X-Total-URLs:
              description: Total number of URLs in sitemap
              type: string
          schema:
            type: string
        "400":
          description: Bad request - missing required parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Generate sitemap
      tags:
      - Sitemap
  /api/sitemap/history:
    get:
      consumes:
      - application/json
      description: Get history of all sitemap generations
      produces:
      - application/json
      responses:
        "200":
          description: List of sitemap generation history
          schema:
            items:
              items:
                $ref: '#/definitions/sitemap.HistoryResponse'
              type: array
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get sitemap generation history
      tags:
      - Sitemap
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"